requirements.txt
humanize==4.9.0
fastapi==0.109.0
Jinja2==3.1.4
jinja2-fragments==1.2.1
requests==2.31.0
requests-toolbelt==1.0.0
pydantic==2.9.2
uvicorn==0.31.1
PyYAML==6.0.1
GitPython==3.1.43
# okta-jwt==1.3.5
